Fix entries not written to the cache
The global themes can have dependencies that are other store items.
Internally the cache has a timer after which the cache will be written.
But in this case we exited the process before this was done.

Test Plan:
Delete `~/.local/share/knewstuff3/ksplash.knsregistry`
to make sure you do not have any noise.

Run:
/path/to/libexec/kf5/kpackagehandlers/knshandler "kns://ksplash.knsrc/api.kde-look.org/1465734/"

Check the file again and make sure there is now an entry present, the
entry should also show up when opening the KSplash KCM and filtering
for installed entries

BUG: 414570
FIXED-IN: 5.79